third_party/WebKit/LayoutTests/dom/domparsing/domparser-document-origin.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test (function() {
	assert_equals((new DOMParser).parseFromString("", "text/html").origin, self.origin);
}, "This test checks that documents created from DOMParser inherit origin");
</script>
